+++ packages/plugin-mock/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @doerjs/plugin-mock
 
+## 0.0.10
+
+### Patch Changes
+
+- Updated dependencies
+  - @doerjs/utils@0.0.10
+
 ## 0.0.9
 
 ### Patch Changes